[ADDED] #165960 Borrar intentos en el árbol de actividades

+++ modules/lms/calification/calification-zero-to-n-web/src/main/java/com/ted/lms/calification/zero/to/n/ZeroToNCalificationType.java
@@ -80,7 +80,7 @@ public class ZeroToNCalificationType extends BaseCalificationType{
 	public String getSuffix() {
 		String suffix = "";
 		if(this.maxValue > 0) {
-			suffix = "/" + this.maxValue;
+			suffix = "/" + (int)this.maxValue;
 		}
 		return suffix;
 	}
